Added more documentation
As the project needs more documentation and the README file itself is not as full, this commit is adding many different files that help the project.

It is adding a CITATION.cff file that allows the user to directly use the citation feature in GitHub and does not need to set up the citation themselves.

Added a CONTRIBUTING.md file that writes down many things that people need to know, when they want to contribute to the project.

Edited extension.json where `special` is replaced with `custom` as `custom` sounds more formal and more fitting to the project.

Edited README.md and replaced `special` with `custom` as `custom` sounds more formal and more fitting to the project.
- Added more information about what is generated by the MultiCodeBlock project.
- Added the Benefits about why the MultiCodeBlock extension is useful.
- Added the full Apache-2.0 license at the bottom of the file.