Change version to 0.1.1
Change Version to 0.1.1 and replaced the screenshot with a screenshot of the specific version.

The Version-Change is made in `extension.json` (for MediaWiki itself to show the version directly) and in `CITATION.cff` (for the citation system in GitHub).

The version change is made as some things have changed, but not things that are so drastic that it should be a change in the minor number.