Replaced the replaceLang function
Replaced the `replaceLang` function in a different file (from MultiCodeBlock.php -> getData.php) and added error checks.

The function didn't have a check, if the language exists and threw an error as well as returned just null and that was behaviour that does not help for special languages where there is no highlighting.